debugging

+++ peoplehub/user_auth/views.py
@@ -5,12 +5,14 @@ from django.contrib.auth import authenticate
 from rest_framework.response import Response
 from rest_framework import status
 from .serializers import RegisterSerializer
+from rest_framework.permissions import AllowAny
 
 # Create your views here.
 
 class LoginAPIView(APIView):
     authentication_classes = []
-    permission_classes = []
+    permission_classes = [AllowAny]
+
     def post(self,request):
         username = request.data.get('username')
         password = request.data.get('password')
@@ -24,16 +26,18 @@ class LoginAPIView(APIView):
     
 class UserRegistration(APIView):
     authentication_classes = []
-    permission_classes = []
+    permission_classes = [AllowAny]
+
     def post(self,request):
         serializer = RegisterSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
-        return Response({"messsage":"User created successfully"},status=status.HTTP_201_CREATED)
+        return Response({"message":"User created successfully"},status=status.HTTP_201_CREATED)
     
 class LogoutAPIView(APIView):
 
-    def post(self,request):
+    def post(self, request):
+      if request.auth is not None:
         request.auth.delete()
-        return Response({"message":"logout Successfull"})
+      return Response({"message": "Logged out successfully."}, status=200)
